chore: move js tests to src/ts/tests
